create alert (hopefully) works
