+++ internal/pkg/postgres/alerts.go
@@ -2,25 +2,58 @@ package postgres
 
 import (
 	"context"
+	"fmt"
+	"time"
 
-	"github.com/byuoitav/smee/internal/smee"
+	"github.com/jackc/pgx/v4"
 )
 
-func (c *Client) CreateAlert(ctx context.Context, alert smee.Alert) (smee.Issue, error) {
-	// get the current issue for the room
-	// if exists, create alert for that issue
-	// if not, create new issue and alert
-	return smee.Issue{}, nil
+type alert struct {
+	ID            int
+	IssueID       int
+	CouchRoomID   string
+	CouchDeviceID string
+	AlertType     string
+	StartTime     time.Time
+	EndTime       *time.Time
 }
 
-func (c *Client) CloseAlert(ctx context.Context, issueID, alertID string) (smee.Issue, error) {
-	return smee.Issue{}, nil
+func (c *Client) createAlert(ctx context.Context, tx pgx.Tx, a alert) (alert, error) {
+	err := tx.QueryRow(ctx,
+		"INSERT INTO alerts (issue_id, couch_room_id, couch_device_id, alert_type, start_time) VALUES ($1, $2, $3, $4, $5) RETURNING id",
+		a.IssueID, a.CouchRoomID, a.CouchDeviceID, a.AlertType, a.StartTime).Scan(&a.ID)
+	if err != nil {
+		return alert{}, fmt.Errorf("unable to query/scan: %w", err)
+	}
+
+	return a, nil
 }
 
-func (c *Client) LinkIncident(ctx context.Context, issueID string, inc smee.Incident) (smee.Issue, error) {
-	return smee.Issue{}, nil
-}
+func (c *Client) alerts(ctx context.Context, tx pgx.Tx, issueID int) ([]alert, error) {
+	var alerts []alert
+	var a alert
 
-func (c *Client) AddIssueEvents(ctx context.Context, issueID string, events ...smee.IssueEvent) error {
-	return nil
+	_, err := tx.QueryFunc(ctx,
+		"SELECT * FROM alerts WHERE issue_id = $1",
+		[]interface{}{issueID},
+		[]interface{}{&a.ID, &a.IssueID, &a.CouchRoomID, &a.CouchDeviceID, &a.AlertType, &a.StartTime, &a.EndTime},
+		func(pgx.QueryFuncRow) error {
+			alerts = append(alerts, alert{
+				ID:            a.ID,
+				IssueID:       a.IssueID,
+				CouchRoomID:   a.CouchRoomID,
+				CouchDeviceID: a.CouchDeviceID,
+				AlertType:     a.AlertType,
+				StartTime:     a.StartTime,
+				EndTime:       a.EndTime,
+			})
+
+			return nil
+		},
+	)
+	if err != nil {
+		return nil, fmt.Errorf("unable to queryFunc: %w", err)
+	}
+
+	return alerts, nil
 }
